requirements.txt
# Generated with: uv pip freeze
django>=5.0.0
djangorestframework>=3.15.0
django-cors-headers>=4.3.0
django-filter>=24.0
drf-spectacular>=0.27.0
python-dotenv>=1.0.0
psycopg[binary]>=3.1.12
gunicorn>=21.2.0
django-allauth
